+++ src/components/mainnav/Mainnav.js
@@ -1,15 +1,15 @@
 import React, {useContext} from 'react';
 import './Mainnav.css';
-import {useHistory, Link} from "react-router-dom";
-import {AuthContext} from "../../context/AuthContext";
+import {Link} from "react-router-dom";
+
 import {ReactComponent as Logo} from "../../assets/icons/logo-weather-heights.svg";
 import Button from "../button/Button";
 import {ReactComponent as Search} from "../../assets/icons/search.svg";
 
 
-function Mainnav () {
-    const {isAuthenticated, userLogoutFunction, email} = useContext(AuthContext);
-    const history = useHistory();
+function Mainnav({children}) {
+
+
     return (
         <div className="outer-container main-nav-background">
             <div className="outer-row">
@@ -19,15 +19,7 @@ function Mainnav () {
                         </span>
                 </div>
                 <div className="mid-nav">
-                    <ul className="outer-row">
-                        <li><Link to="/"> France</Link></li>
-                        <li><Link to="/"> Regions</Link></li>
-                        <li><Link to="/"> Departments </Link></li>
-                        {isAuthenticated &&
-                            <li><Link to="/"> MyLocations </Link></li>
-                        }
-                        <li><Link to="/"> Cities </Link></li>
-                    </ul>
+                    {children}
                 </div>
                 <div className="right-nav">
                     <button>
