+++ src/index.js
@@ -7,13 +7,13 @@ import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <React.StrictMode>
-        <Router>
+    <Router>
+        <React.StrictMode>
             <AuthContextProvider>
                 <App/>
             </AuthContextProvider>
-        </Router>
-    </React.StrictMode>
+        </React.StrictMode>
+    </Router>
 );
