+++ src/pages/searchcity/Searchcity.css
@@ -7,8 +7,8 @@
     background: var(--main-body);
     width: 100%;
     padding: 0 10px;
-    margin: 0 20px;
-    /*height: 71px;*/
+    margin: 10px 20px;
+  height: 60px;
     border: solid 2px var(--bordertile);
     border-radius: 8px;
 }
